#322: List of tasks - deleting task should not be available when task is assigned to at least one plan
* deleting task which is assigned to a plan is no longer possible. AlertDialog is displayed. Test added.

* unnecessary test removal

* creation of DialogUserNotifier, getPlansWithThisTask() moved to PlanTemplateRepository and some other minor changes/fixes

+++ Friendly-plans/app/src/main/java/pg/autyzm/friendly_plans/manager_app/view/task_list/TaskListActivity.java
@@ -5,18 +5,28 @@ import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
+import android.text.TextUtils;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import database.entities.PlanTemplate;
+import database.repository.PlanTemplateRepository;
 import database.repository.TaskTemplateRepository;
 import javax.inject.Inject;
 import pg.autyzm.friendly_plans.ActivityProperties;
 import pg.autyzm.friendly_plans.App;
 import pg.autyzm.friendly_plans.R;
 import pg.autyzm.friendly_plans.manager_app.view.task_create.TaskCreateActivity;
+import pg.autyzm.friendly_plans.notifications.DialogUserNotifier;
 import pg.autyzm.friendly_plans.notifications.ToastUserNotifier;
 
 public class TaskListActivity extends AppCompatActivity {
 
     @Inject
     TaskTemplateRepository taskTemplateRepository;
+    @Inject
+    PlanTemplateRepository planTemplateRepository;
     @Inject
     ToastUserNotifier toastUserNotifier;
 
@@ -27,11 +37,18 @@ public class TaskListActivity extends AppCompatActivity {
 
                 @Override
                 public void onRemoveTaskClick(int position){
-                    taskTemplateRepository.delete(taskListAdapter.getTaskItem(position).getId());
-                    toastUserNotifier.displayNotifications(
-                            R.string.task_removed_message,
-                            getApplicationContext());
-                    taskListAdapter.setTaskItems(taskTemplateRepository.getAll());
+                    List<PlanTemplate> relatedPlans = planTemplateRepository.getPlansWithThisTask(
+                            taskListAdapter.getTaskItem(position).getId());
+                    if(relatedPlans.isEmpty()) {
+                        taskTemplateRepository.delete(taskListAdapter.getTaskItem(position).getId());
+                        toastUserNotifier.displayNotifications(
+                                R.string.task_removed_message,
+                                getApplicationContext());
+                        taskListAdapter.setTaskItems(taskTemplateRepository.getAll());
+                    }
+                    else{
+                        displayTaskCannotBeRemovedAlert(relatedPlans);
+                    }
                 }
 
                 @Override
@@ -67,5 +84,22 @@ public class TaskListActivity extends AppCompatActivity {
         recyclerView.setLayoutManager(new LinearLayoutManager(this));
         taskListAdapter = new TaskRecyclerViewAdapter(taskItemClickListener);
         recyclerView.setAdapter(taskListAdapter);
+    }
+
+    private void displayTaskCannotBeRemovedAlert(List<PlanTemplate> relatedPlans) {
+        List<String> relatedPlansNames = new ArrayList<>();
+        for (PlanTemplate plan : relatedPlans){
+            relatedPlansNames.add(plan.getName());
+        }
+        String relatedPlansNamesJoined = TextUtils.join(", " , relatedPlansNames);
+
+        DialogUserNotifier dialog = new DialogUserNotifier(
+                TaskListActivity.this,
+                getResources().getString(R.string.task_cannot_be_removed),
+                getResources().getString(R.string.task_cannot_be_removed_message)
+                        + " " + relatedPlansNamesJoined + ".",
+                getResources().getString(R.string.task_cannot_be_removed_dialog_close_button_text)
+                );
+        dialog.showDialog();
     }
 }
